mac80211: add a general fallback TXQ
In a number of cases, like management frames, mac80211 will not
put the frame on any TXQ but immediately TX it to the driver.
It'd be nicer to be able to use TXQs for all frames, so add a
"fallback" TXQ. This will serve as a container to be able to
remove all the non-TXQ queueing in mac80211.

For ath9k/ath10k, extend the code that already handles the STA
mgmt TXQ by code to handle the per-HW fallback queue. Since that
may now carry fragmented frames, extend the logic a bit.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>